requirements.txt
torch~=2.2.2
torchvision~=0.17.2
facenet-pytorch
pillow~=10.2.0
matplotlib~=3.10.0
numpy
pandas~=2.2.3